chainntnfs+queue: move ConcurrentQueue to own package 'queue'
