+++ queue/queue_test.go
@@ -1,13 +1,13 @@
-package chainntnfs_test
+package queue_test
 
 import (
 	"testing"
 
-	"github.com/lightningnetwork/lnd/chainntnfs"
+	"github.com/lightningnetwork/lnd/queue"
 )
 
 func TestConcurrentQueue(t *testing.T) {
-	queue := chainntnfs.NewConcurrentQueue(100)
+	queue := queue.NewConcurrentQueue(100)
 	queue.Start()
 	defer queue.Stop()
